add: Added Summer Cocktail
Related to request #375

SC used in following script are not existing (yet) in the list of effect
[sc_start SC_DEF_RATE,1800000,10;]
[sc_start SC_INCATKRATE,1800000,5;]
[sc_start SC_JEXPBOOST,1800000,15;]
[sc_start SC_INCASPDRATE,2700000,10;]
[sc_start SC_CASTRATE,2700000,-10;]
[sc_start SC_INCMATKRATE,1800000,5;]
[sc_start SC_LIFEINSURANCE,1800000,0; sc_start SC_EXPBOOST,1800000,10;]
[specialeffect2 320; sc_start4 SC_L_LIFEPOTION,600000,-3,10,0,0; sc_start SC_INCFLEE,600000,30;]
[sc_start SC_MDEF_RATE,1800000,10;]

Not worth adding them just for the purpose of managing those cocktails, if later on those SC are more often used, then it could be interesting to define them.

The way ATK bonus is managed needs to be re-investigated (Impositio, Drum, Volcano, Venatu Cocktail, ...), it seems it should managed through the weapon attack instead.